requirements.txt
amqp==2.5.0
asn1crypto==0.24.0
backports.csv==1.0.7
beautifulsoup4==4.7.1
billiard==3.6.0.0
bs4==0.0.1
celery==4.3.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.1.7
django-crispy-forms==1.7.2
django-formtools==2.1
django-haystack==2.8.1
django-import-export==1.2.0
django-redis==4.10.0
django-reversion==3.0.3
django-tinymce==2.8.0
et-xmlfile==1.0.1
fake-useragent==0.1.11
future==0.17.1
httplib2==0.12.3
idna==2.8
itsdangerous==1.1.0
jdcal==1.4.1
jieba==0.39
Jinja2==2.10
kombu==4.6.0
lxml==4.3.2
Markdown==3.1
MarkupSafe==1.1.1
odfpy==1.4.0
openpyxl==2.6.2
Pillow==6.0.0
pycparser==2.19
pycryptodomex==3.7.2
PyMySQL==0.9.2
pyOpenSSL==19.0.0
python-alipay-sdk==1.10.0
pytz==2019.1
PyYAML==5.1
redis==3.2.1
requests==2.21.0
six==1.12.0
soupsieve==1.9.1
tablib==0.13.0
urllib3==1.24.3
vine==1.3.0
Whoosh==2.7.4
xlrd==1.2.0
xlwt==1.3.0
